* [#1933] fix(multidb): fix some methods to call the correct db ** fix some  leaks database connections when configured for multiple databases * refactor(JPAModelLoader): move JPAModelLoader class in its own file
